Fix redbean SQLite for closing db multiple times (explicitly or using _gc)
The loop that releases and cleans all vm in cleanupdb is likely not
needed, as each of the statement/vm is either closed explicitly or is
finalized by its __gc handler (which is done in cleanupvm). The loop
also relies on all values to being tables (so they can be all removed
from cleanupvm) and restarts the search from the beginning, so if even
one element is left, the loop can't be completed.

I've observed cases under load (with a large number of statements/vm
created) when the value retrieved in cleanupvm is not a table, which
led to SIGSEGV with the following stacktrace:

70000003de20 5df71a getgeneric+26
70000003fac0 5dfc7f luaH_get+111
70000003faf0 5e06c8 luaH_set+40
70000003fb20 5c5bd7 aux_rawset+55
70000003fb50 5c70cb lua_rawset+27
70000003fb60 4fa8e7 cleanupvm+71
70000003fb80 4fa988 cleanupdb+88
70000003fbc0 4fe899 db_gc+41

This patch adds a check for the type of the object (to avoid the
SIGSEGV) and removes the loop (as described earlier), since adding
a check is not sufficient, as it prevents removal of all elements
and the loop from finishing as the result.

The patch also switches to using close_v2, as it is intended for use
with garbage collected languages where the order in which destructors
are called is arbitrary, as is the case here.